+++ src/command_modules/azure-cli-iot/azure/cli/command_modules/iot/_constants.py
@@ -2,5 +2,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-import pkg_resources
-pkg_resources.declare_namespace(__name__)
+
+certificate_help = """For a detailed explanation of CA certificates in Azure IoT Hub,
+                      see https://docs.microsoft.com/en-us/azure/iot-hub/iot-hub-x509ca-overview """