+++ src/command_modules/azure-cli-reservations/azure/cli/command_modules/reservations/_exception_handler.py
@@ -2,5 +2,16 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-import pkg_resources
-pkg_resources.declare_namespace(__name__)
+
+from azure.cli.core.util import CLIError
+
+
+def reservations_exception_handler(ex):
+    from azure.mgmt.reservations.models import Error
+    if isinstance(ex, Error):
+        message = ex.error.error.message
+        raise CLIError(message)
+    else:
+        import sys
+        from six import reraise
+        reraise(*sys.exc_info())