+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/tests/test_role_commands_thru_mock.py
@@ -8,12 +8,14 @@ import tempfile
 import unittest
 import uuid
 import mock
-
+from azure.cli.core.util import CLIError
 from azure.mgmt.authorization.models import RoleDefinition, RoleDefinitionProperties
-from azure.graphrbac.models import Application, ServicePrincipal
+from azure.graphrbac.models import Application, ServicePrincipal, GraphErrorException
 from azure.cli.command_modules.role.custom import (create_role_definition,
                                                    update_role_definition,
-                                                   create_service_principal_for_rbac)
+                                                   create_service_principal_for_rbac,
+                                                   reset_service_principal_credential,
+                                                   _try_x509_pem)
 
 # pylint: disable=line-too-long
 
@@ -157,10 +159,133 @@ class TestRoleMocked(unittest.TestCase):
         self.assertTrue(logger_mock.warning.called)  # we should warn 'years' will be dropped
         self.assertTrue(faked_graph_client.applications.create.called)
 
+    @mock.patch('azure.cli.command_modules.role.custom._graph_client_factory', autospec=True)
+    def test_reset_credentials_password(self, graph_client_mock):
+        patch_invoked = [False]  # to be used in a nested function below, array type is needed to get scoping work
+        test_object_id = 'app_object_id'
+        test_pwd = 'verySecret'
+        name = 'http://mysp'
 
-class FakedResponse(object):  # pylint: disable=too-few-public-methods
-    def __init__(self, status_code):
-        self.status_code = status_code
+        def test_patch(id, param):
+            patch_invoked[0] = True
+            self.assertEqual(id, test_object_id)
+            self.assertEqual(1, len(param.password_credentials))
+            self.assertEqual(param.password_credentials[0].value, test_pwd)
+
+        faked_graph_client = mock.MagicMock()
+        sp_object = mock.MagicMock()
+        sp_object.app_id = 'app_id'
+        app_object = mock.MagicMock()
+        app_object.object_id = test_object_id
+
+        graph_client_mock.return_value = faked_graph_client
+        faked_graph_client.service_principals.list.return_value = [sp_object]
+        faked_graph_client.applications.list.return_value = [app_object]
+        faked_graph_client.applications.get.side_effect = [app_object]
+        faked_graph_client.applications.patch = test_patch
+        faked_graph_client.applications.list_password_credentials.side_effect = [ValueError('should not invoke')]
+
+        # action
+        reset_service_principal_credential(name, test_pwd, append=False)
+
+        # assert
+        self.assertTrue(patch_invoked[0])
+
+    @mock.patch('azure.cli.command_modules.role.custom._graph_client_factory', autospec=True)
+    def test_reset_credentials_certificate_append_option(self, graph_client_mock):
+        patch_invoked = [False]  # to be used in a nested function below, array type is needed to get scoping work
+        test_object_id = 'app_object_id'
+        test_cert = _try_x509_pem('\n'.join(['-----BEGIN CERTIFICATE-----',
+                                             'MIICoTCCAYkCAgPoMA0GCSqGSIb3DQEBBQUAMBQxEjAQBgNVBAMMCUNMSS1Mb2dp',
+                                             'bjAiGA8yMDE3MTExMzIxMjQyMloYDzIwMTgxMTEzMjEyNDI0WjAUMRIwEAYDVQQD',
+                                             'DAlDTEktTG9naW4wggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCtyRA6',
+                                             'mbUtByQBODMlp3r5fGpnYCfRhAzp2U29KRVTOQK1ntMIo3FWR19ceqK6T0UM+BFb',
+                                             'XGn28hGhgz5Y1lrbqyKrAcF10/3y42wmiMyPWjmXJ+WOEKjckKKzPMm2KBsn/ePV',
+                                             'qsr5UwlnHh2rGFR0PF1qjC0IU/SI0UQN0KJKpVp0OB8+lRlIAcsLUTveTXbdFDlp',
+                                             'k5AA8w3TTo7pT5sKNZr3+qv1o4ogDfDEx0bCYtlm4L1HvGer4pbX7q35ucZY9BWq',
+                                             'VjHQ/MjiuAyxZoyY5xVoULMWJupRyMT3wP1Hb+oJ9/tTBZbpNTv1ed9OswCc2W1l',
+                                             'MQzLwE10Ev0LJkhlAgMBAAEwDQYJKoZIhvcNAQEFBQADggEBAF+RP+7uzmf4u4+l',
+                                             'xjS+lXd3O0tqTNOVe8RR+GXl1s2Un6UhwmP3SKs2RNcXkb9+6Q4Zvg7GODK7Z8th',
+                                             't/enpTcvLmPmq2ow1hFGWJk/lONZtVKU2mikTY/ICnQrbhf3WYr1cuf98CRqoG71',
+                                             'ldrjgSsM1Ut7Gee1Tpc2eamRtHTm07AUQhqGnS5wVp6s1HUd43nvu/lVx+j2hjEB',
+                                             'y63BSuD3aSUweVne4roBNcBJjLU1wvYl+3cLgnZ9///3y/C4pKebsKHljkejRaer',
+                                             'nvbPfW9hy7BqMem7t0Qk2D84VzaK+6x9EnnsXy+90nfvTLUSqpU1MjpdWhuWyxDL',
+                                             'p4oYS5Q=',
+                                             '-----END CERTIFICATE-----']))
+        key_id_of_existing_cert = 'existing cert'
+        name = 'http://mysp'
+
+        def test_patch(id, param):
+            patch_invoked[0] = True
+            self.assertEqual(id, test_object_id)
+            self.assertEqual(2, len(param.key_credentials))
+            self.assertTrue(len(param.key_credentials[1].value) > 0)
+            self.assertEqual(param.key_credentials[1].type, 'AsymmetricX509Cert')
+            self.assertEqual(param.key_credentials[1].usage, 'Verify')
+            self.assertEqual(param.key_credentials[0].key_id, key_id_of_existing_cert)
+
+        faked_graph_client = mock.MagicMock()
+        sp_object = mock.MagicMock()
+        sp_object.app_id = 'app_id'
+        app_object = mock.MagicMock()
+        app_object.object_id = test_object_id
+        key_cred = mock.MagicMock()
+        key_cred.key_id = key_id_of_existing_cert
+
+        graph_client_mock.return_value = faked_graph_client
+        faked_graph_client.service_principals.list.return_value = [sp_object]
+        faked_graph_client.applications.list.return_value = [app_object]
+        faked_graph_client.applications.get.side_effect = [app_object]
+        faked_graph_client.applications.patch = test_patch
+        faked_graph_client.applications.list_key_credentials.return_value = [key_cred]
+
+        # action
+        reset_service_principal_credential(name, cert=test_cert, append=True)
+
+        # assert
+        self.assertTrue(patch_invoked[0])
+
+    @mock.patch('azure.cli.command_modules.role.custom._auth_client_factory', autospec=True)
+    @mock.patch('azure.cli.command_modules.role.custom._graph_client_factory', autospec=True)
+    def test_create_for_rbac_failed_with_polished_error_if_due_to_permission(self, graph_client_mock, auth_client_mock):
+        TestRoleMocked._common_rbac_err_polish_test_mock_setup(graph_client_mock, auth_client_mock,
+                                                               'Insufficient privileges to complete the operation',
+                                                               self.subscription_id)
+
+        # action
+        with self.assertRaises(CLIError) as context:
+            create_service_principal_for_rbac('will-fail', skip_assignment=True)
+
+        # assert we handled such error
+        self.assertTrue('https://docs.microsoft.com/en-us/azure/azure-resource-manager/resource-group-create-service-principal-portal' in str(context.exception))
+
+    @mock.patch('azure.cli.command_modules.role.custom._auth_client_factory', autospec=True)
+    @mock.patch('azure.cli.command_modules.role.custom._graph_client_factory', autospec=True)
+    def test_create_for_rbac_failed_with_regular_error(self, graph_client_mock, auth_client_mock):
+        TestRoleMocked._common_rbac_err_polish_test_mock_setup(graph_client_mock, auth_client_mock,
+                                                               'something bad for you',
+                                                               self.subscription_id)
+        # action
+        self.assertRaises(GraphErrorException, create_service_principal_for_rbac, 'will-fail')
+
+    @staticmethod
+    def _common_rbac_err_polish_test_mock_setup(graph_client_mock, auth_client_mock, error_msg, subscription_id):
+        def _test_deserializer(resp_type, response):
+            err = FakedError(error_msg)
+            return err
+
+        faked_role_client = mock.MagicMock()
+        faked_role_client.config.subscription_id = subscription_id
+        auth_client_mock.return_value = faked_role_client
+        faked_graph_client = mock.MagicMock()
+        graph_client_mock.return_value = faked_graph_client
+
+        faked_graph_client.applications.create.side_effect = GraphErrorException(_test_deserializer, None)
+
+
+class FakedError(object):  # pylint: disable=too-few-public-methods
+    def __init__(self, message):
+        self.message = message
 
 
 if __name__ == '__main__':