+++ src/command_modules/azure-cli-storage/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+2.0.20
+++++++
+* Update multiapi storage package dependency to 0.1.7
+
 2.0.19
 ++++++
 * `storage account create`: defaults --sku to 'Standard_RAGRS'